requirements.txt
python-telegram-bot==13.15
environs==9.5.0
redis==4.4.0
vk-api==11.9.9